+++ hashtable.h
@@ -12,14 +12,14 @@ typedef void** Bucket;
 typedef struct HashTable
 {
     Bucket** buckets;
-    int size;
+    unsigned long int size;
     size_t bucketSize;
 } HashTable;
 
 Bucket* initializeBucket(size_t bucketSize);
 int checkBucketHasNext(Bucket* bucket, size_t bucketSize);
 void insertToBucket(Bucket* bucket, void* item, size_t bucketSize);
-HashTable* initializeHashTable(int hashTableSize, size_t bucketSize);
+HashTable* initializeHashTable(unsigned long int hashTableSize, size_t bucketSize);
 unsigned long int hash_function(char* string, int max);
 
 #endif //ERGASIA_1_HASHTABLE_H
